--- client/src/components/Alert/styles.ts
@@ -1,13 +1,0 @@
-import { makeStyles } from "@material-ui/core/styles";
-
-const useStyles = makeStyles((theme) => {
-  return {
-    root: {
-      display: "flex",
-      justifyContent: "center",
-      alignItems: "center",
-    },
-  };
-});
-
-export default useStyles;
