+++ client/src/components/DetailTable.tsx
@@ -1,9 +1,40 @@
 import { useRef, useState } from "react";
 import { Tooltip, Typography } from "@material-ui/core/";
-import CopyToClipBoard from "../CopyToClipBoard/CopyToClipBoard";
+import CopyToClipBoard from "./CopyToClipBoard";
 
 // STYLES
-import useStyles from "./styles";
+import { makeStyles, Theme } from "@material-ui/core/styles";
+
+const useStyles = makeStyles((theme: Theme) => ({
+  container: {
+    padding: theme.spacing(1.2),
+    overflow: "auto",
+    display: "grid",
+    gridTemplateColumns: "1fr ",
+    gap: theme.spacing(2),
+    rowGap: theme.spacing(2),
+    justifyContent: "space-between",
+  },
+  bold: {
+    fontWeight: "bold",
+  },
+  bullet: {
+    cursor: "pointer",
+  },
+  ellipsis: {
+    overflow: "hidden",
+    whiteSpace: "nowrap",
+    textOverflow: "ellipsis",
+  },
+  row: {
+    height: "24px",
+    display: "flex",
+    overflow: "hidden",
+    alignItems: "center",
+  },
+  keys: { minWidth: "130px" },
+  values: {},
+}));
 
 type Trows = {
   rows: { key: string; value: string }[];
