--- client/src/components/SearchBar/styles.ts
@@ -1,9 +1,0 @@
-import { createStyles, Theme, makeStyles } from "@material-ui/core/styles";
-
-export const useStyles = makeStyles((theme: Theme) =>
-  createStyles({
-    searchBar: {
-      width: "100%",
-    },
-  })
-);
