+++ client/src/pages/Details/ProductDetails.tsx
@@ -5,16 +5,125 @@ import AccordionSummary from "@material-ui/core/AccordionSummary";
 import AccordionDetails from "@material-ui/core/AccordionDetails";
 import Typography from "@material-ui/core/Typography";
 import ExpandMoreIcon from "@material-ui/icons/ExpandMore";
-import DetailTable from "../../../components/DetailTable/DetailTable";
+import DetailTable from "../../components/DetailTable";
 import { useParams } from "react-router";
-import getProductInfo from "../../../api/getProductInfo";
+import getProductInfo from "../../api/getProductInfo";
 import { useQuery } from "@tanstack/react-query";
-import { useStore } from "../../DrawerAppbar/DrawerAppbar";
-import CustomAlert from "../../../components/Alert/CustomAlert";
+import { useStore } from "../DrawerAppbar";
+import CustomAlert from "../../components/CustomAlert";
 import _ from "lodash";
 
 // STYLES
-import useStyles from "./styles";
+import { makeStyles, Theme } from "@material-ui/core/styles";
+
+const useStyles = makeStyles((theme: Theme) => ({
+  gridContainer: {
+    background: theme.palette.background.paper,
+    height: "calc(100vh - 115px)",
+    width: "100%",
+    display: "flex",
+    justifyContent: "space-between",
+    gap: theme.spacing(1),
+    flexWrap: "nowrap",
+  },
+  leftBox: {
+    background: theme.palette.background.default,
+    height: "100%",
+    position: "relative",
+  },
+  rightBox: {
+    padding: theme.spacing(1, 0.5),
+    position: "relative",
+    overflow: "auto",
+  },
+
+  imagesContainer: {
+    height: "inherit",
+  },
+  carousel: {
+    overflow: "auto",
+    height: "100%",
+    width: "100%",
+  },
+
+  img: {
+    cursor: "pointer",
+    width: "100%",
+    objectFit: "scale-down",
+    maxHeight: "170px",
+    paddingBottom: theme.spacing(0.25),
+    "&:hover": {
+      paddingBottom: theme.spacing(0),
+      borderBottom: `2px solid ${theme.palette.primary.main}`,
+    },
+  },
+  mainImgGrid: {
+    display: "flex",
+    justifyContent: "center",
+    alignItems: "center",
+    width: "100%",
+    height: "100%",
+    background: theme.palette.background.default,
+  },
+  mainImage: {
+    objectFit: "scale-down",
+    height: "100%",
+    width: "100%",
+    padding: theme.spacing(0.3),
+    flexGrow: 1,
+  },
+  accordionBox: {
+    width: "100%",
+  },
+  header: {
+    display: "grid",
+    gridTemplateColumns: "1fr",
+    gap: theme.spacing(2),
+  },
+  divider: {
+    marginBottom: theme.spacing(1),
+  },
+  details: {
+    display: "grid",
+    gridTemplateColumns: "1fr",
+    padding: theme.spacing(1, 2, 2, 2),
+  },
+  extraStyles: {
+    "& [class*='row']": {
+      gap: theme.spacing(2),
+    },
+  },
+  idBox: {
+    display: "flex",
+    justifyContent: "space-between",
+  },
+  productTitle: {
+    padding: theme.spacing(0, 0, 2, 0),
+  },
+  productDesc: {
+    display: "grid",
+    gap: theme.spacing(2),
+  },
+  center: {
+    display: "block",
+    position: "absolute",
+    top: "50%",
+    left: "50%",
+  },
+  accordionTitle: {
+    color: theme.palette.primary.main,
+    "&:hover": {
+      marginBottom: theme.spacing(-1 / 4),
+      borderBottom: `2px solid ${theme.palette.primary.main}`,
+    },
+  },
+  icon: {
+    color: theme.palette.primary.main,
+  },
+  accordion: {
+    boxShadow: "none",
+  },
+}));
 
 type Tproduct = {
   id: string;
