+++ client/src/pages/DrawerAppbar.tsx
@@ -18,15 +18,133 @@ import {
 } from "@fortawesome/free-solid-svg-icons";
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import { BrowserRouter, Link, Route, Switch } from "react-router-dom";
-import CatalogExplorer from "../CatalogExplorer/CatalogExplorer";
-import ProductsList from "../ProductsList/ProductsList";
-import ProductDetails from "../Details/ProductDetails/ProductDetails";
-import AddProducts from "../../components/AddProducts/AddProducts";
+import CatalogExplorer from "./CatalogExplorer";
+import ProductsList from "./ProductsList";
+import ProductDetails from "./Details/ProductDetails";
+import AddProducts from "../components/AddProducts";
 import { create } from "zustand";
 import { persist } from "zustand/middleware";
-import CustomNavBar from "../../components/CustomNavBar/CustomNavBar";
-import queryClientConfig from "../../config/queryClientConfig";
-import useStyles from "./styles";
+import CustomNavBar from "../components/CustomNavBar";
+import queryClientConfig from "../config/queryClientConfig";
+
+import { createStyles, makeStyles, Theme } from "@material-ui/core/styles";
+
+const drawerWidth = 240;
+const drawerWidthMin = 70;
+
+const useStyles = makeStyles((theme: Theme) =>
+  createStyles({
+    root: {
+      display: "flex",
+      "& .Mui-selected": {
+        backgroundColor: theme.palette.primary.light,
+      },
+    },
+
+    appBar: {
+      boxShadow: "none",
+      borderBottom: `${theme.spacing(1) / 8}px solid ${
+        theme.palette.action.focus
+      }`,
+      zIndex: theme.zIndex.drawer - 1,
+      width: `calc(100% - ${drawerWidthMin}px)`,
+      transition: theme.transitions.create(["width", "margin"], {
+        easing: theme.transitions.easing.sharp,
+        duration: theme.transitions.duration.leavingScreen,
+      }),
+      backgroundColor: theme.palette.background.paper,
+      color: theme.palette.getContrastText(theme.palette.background.paper),
+    },
+    appBarShift: {
+      marginLeft: drawerWidth,
+      width: `calc(100% - ${drawerWidth}px)`,
+      transition: theme.transitions.create(["width", "margin"], {
+        easing: theme.transitions.easing.sharp,
+        duration: theme.transitions.duration.enteringScreen,
+      }),
+    },
+    hide: {
+      display: "none",
+    },
+    drawer: {
+      width: drawerWidth,
+      flexShrink: 0,
+      whiteSpace: "nowrap",
+    },
+    drawerOpen: {
+      width: drawerWidth,
+      transition: theme.transitions.create("width", {
+        easing: theme.transitions.easing.sharp,
+        duration: theme.transitions.duration.enteringScreen,
+      }),
+    },
+    drawerClose: {
+      transition: theme.transitions.create("width", {
+        easing: theme.transitions.easing.sharp,
+        duration: theme.transitions.duration.leavingScreen,
+      }),
+      overflowX: "hidden",
+      width: theme.spacing(7) + 1,
+      [theme.breakpoints.up("sm")]: {
+        width: theme.spacing(9) + 1,
+      },
+    },
+    toolbar: {
+      display: "flex",
+      alignItems: "center",
+      justifyContent: "flex-end",
+      padding: theme.spacing(0, 1),
+      // necessary for content to be below app bar
+      ...theme.mixins.toolbar,
+    },
+    content: {
+      flexGrow: 1,
+      padding: theme.spacing(2),
+      minHeight: "100vh",
+    },
+    buttonList: {
+      height: "100vh",
+      display: "flex",
+      flexDirection: "column",
+      padding: theme.spacing(1, 1),
+    },
+    buttonStyle: {
+      borderRadius: theme.shape.borderRadius,
+      marginBottom: theme.spacing(1),
+      color: theme.palette.grey[600],
+    },
+    buttonStyleSelected: {
+      color: theme.palette.primary.main,
+      borderRadius: theme.shape.borderRadius,
+      marginBottom: theme.spacing(1),
+      "&:hover": {
+        backgroundColor: `${theme.palette.primary.light} !important`,
+      },
+    },
+    iconSelected: {
+      color: theme.palette.primary.main,
+    },
+    textButton: {
+      letterSpacing: "0.04em",
+    },
+    drawerHeader: {
+      borderRadius: theme.shape.borderRadius,
+      marginBottom: theme.spacing(1),
+    },
+    drawerTitle: {
+      fontSize: "18px",
+      letterSpacing: "0.04em",
+    },
+    flexGrow: {
+      flexGrow: 1,
+    },
+    link: {
+      color: theme.palette.text.primary,
+      textDecoration: "none",
+    },
+  })
+);
+
 const drawerButtons = [
   {
     text: "Data Explorer",
